prefetch filter data on site load
