+++ src/types/Filters.ts
@@ -3,3 +3,8 @@ export type Filters = {
   Attributes: string[];
   Semesters: string[];
 };
+
+export type FilterData = {
+  id: number;
+  code: string;
+};
